flask/requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-Mail==0.7.6
Flask-SQLAlchemy==0.16
Flask-Testing==0.4.2
Flask-WTF==0.8.3
Jinja2==2.6
SQLAlchemy==0.8.0
Tempita==0.5.1
WTForms==1.0.3
Werkzeug==0.8.3
blinker==1.2
decorator==3.4.0
ipython==0.13.1
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
py-healthcheck
